feat: add runBuild helper
